cdylib in its own file, NonNull
