+++ src/lib.rs
@@ -116,6 +116,9 @@ pub use pixel_format::*;
 mod palette;
 pub use palette::*;
 
+mod cdylib;
+pub use cdylib::*;
+
 /// Grabs up the data from a null terminated string pointer.
 unsafe fn gather_string(ptr: *const c_char) -> String {
   let len = SDL_strlen(ptr);
@@ -364,14 +367,12 @@ impl SDLToken {
   /// Attempts to load the named dynamic library into the program.
   pub fn load_cdylib<'sdl>(&'sdl self, name: &str) -> Result<CDyLib<'sdl>, String> {
     let name_null: Vec<u8> = name.bytes().chain(Some(0)).collect();
-    let ptr = unsafe { SDL_LoadObject(name_null.as_ptr() as *const c_char) };
-    if ptr.is_null() {
-      Err(get_error())
-    } else {
-      Ok(CDyLib {
-        ptr,
+    match NonNull::new(unsafe { SDL_LoadObject(name_null.as_ptr() as *const c_char) }) {
+      Some(nn) => Ok(CDyLib {
+        nn,
         _marker: PhantomData,
-      })
+      }),
+      None => Err(get_error()),
     }
   }
 
@@ -959,49 +960,5 @@ impl From<Rect> for SDL_Rect {
       w: other.w,
       h: other.h,
     }
-  }
-}
-
-/// Handle to a C ABI dynamic library that has been loaded.
-///
-/// You can make your own libs that will work with this using the `cdylib` crate
-/// type
-/// ([here](https://rust-embedded.github.io/book/interoperability/rust-with-c.html)
-/// is a short tutorial).
-///
-/// Do not attempt to mix this with the `dylib` crate type. That's a crate type
-/// you should not use, it's basically for `rustc` internal usage only.
-#[derive(Debug)]
-#[repr(transparent)]
-pub struct CDyLib<'sdl> {
-  ptr: *mut c_void,
-  _marker: PhantomData<&'sdl SDLToken>,
-}
-impl<'sdl> Drop for CDyLib<'sdl> {
-  fn drop(&mut self) {
-    unsafe { SDL_UnloadObject(self.ptr) }
-  }
-}
-impl<'sdl> CDyLib<'sdl> {
-  /// Attempts to look up a function by name, getting its pointer.
-  ///
-  /// Once this function returns, you will have to
-  /// [transmute](core::mem::transmute) the optional NonNull value you get into
-  /// an optional `fn` value of some sort.
-  ///
-  /// You _probably_ want to transmute it into `Option<unsafe extern "C"
-  /// fn(INPUTS) -> OUTPUTS>`, but it's possible that you might need to use some
-  /// other ABI for example. This whole thing is obviously not at all safe. You
-  /// absolutely must get the `fn` type correct when doing this `transmute`.
-  ///
-  /// # Safety
-  ///
-  /// * The returned value _does not_ have a lifetime linking it back to this
-  ///   shared library. Making sure that the function pointer is not used after
-  ///   this library unloads is up to you.
-  pub unsafe fn find_function(&self, name: &str) -> Option<NonNull<c_void>> {
-    let name_null: Vec<u8> = name.bytes().chain(Some(0)).collect();
-    let name_ptr: *const c_char = name_null.as_ptr() as *const c_char;
-    NonNull::new(SDL_LoadFunction(self.ptr, name_ptr))
   }
 }
